phoneCall & Location modal components created
