+++ client/src/pages/EventTypePage/SoloEventPage/SoloEventPage.js
@@ -11,6 +11,8 @@ import Header from '../../../components/Header/Header';
 import FormSubmitControls from './FormSubmitControls/FormSubmitControls';
 import EventTypeHeader from '../EventTypeHeader/EventTypeHeader';
 import RadioColorList from './RadioColorList/RadioColorList';
+import PhoneCallModal from './Modals/PhoneCallModal';
+import LocationModal from './Modals/LocationModal';
 
 class SoloEventPage extends Component {
     constructor(props){
@@ -170,93 +172,3 @@ class SoloEventPage extends Component {
 };
 
 export default withRouter(SoloEventPage);
-
-
-const LocationModal = ({ handleLocationUpdate, handleLocationModal, handleUserInput }) => (
-    <Box className="locationModal">
-        <Box boxShadow={3} className="locationModal__window">
-            <Box className="locationModal__window--header">
-                Edit Location
-            </Box>
-            <Box className="locationModal__window--col">
-                <Box className="locationModal__window--col--selected">
-                    <LocationOnIcon className="soloEvent__form--location--settings--dropdown--icon"/>
-                    <Typography variant="body2">
-                        In-person meeting
-                    </Typography>
-                </Box>
-                <input 
-                    onChange={handleUserInput}
-                    name="eventLocation" 
-                    className="locationModal__window--col--input" 
-                    autoComplete="off"/>
-            </Box>
-            <Box className="locationModal__window--footer">
-                <Button 
-                    onClick={handleLocationUpdate}
-                    className="locationModal__window--btn locationModal__window--btn--update"
-                >Update</Button>
-                <Button 
-                    onClick={handleLocationModal}
-                    className="locationModal__window--btn locationModal__window--btn--cancel"
-                >Cancel</Button>
-            </Box>
-        </Box>
-    </Box>
-);
-
-const PhoneCallModal = ({ handleLocationUpdate, handlePhoneCallModal, handleUserInput, eventLocation }) => (
-    <Box className="phoneCallModal">
-        <Box boxShadow={3} className="phoneCallModal__window">
-            <Box className="phoneCallModal__window--header">
-                Edit Location
-            </Box>
-            <Box className="phoneCallModal__window--col">
-                <Box className="phoneCallModal__window--col--selected">
-                    <CallIcon className="soloEvent__form--location--settings--dropdown--icon"/>
-                    <Typography variant="body2">
-                        Phone Call
-                    </Typography>
-                </Box>
-                <Box className="phoneCallModal__window--radios">
-                    <Box className="phoneCallModal__window--radio--wrap">
-                        <Radio
-                            onChange={handleUserInput}
-                            checked={eventLocation === 'phone call: (I will call invitee)'}
-                            name="eventLocation"
-                            className="phoneCallModal__window--radio"
-                            value="phone call: (I will call invitee)"
-                        />
-                        <Typography className="phoneCallModal__window--radio--label" variant="body2">
-                            I will call my invitee<br/>
-                            <span className="phoneCallModal__window--radio--label--span">Calendly will require your invitee’s phone number before scheduling.</span>
-                        </Typography>
-                    </Box>
-                    <Box className="phoneCallModal__window--radio--wrap">
-                        <Radio
-                            onChange={handleUserInput}
-                            checked={eventLocation === 'phone call: (Invitee should call me)'}
-                            name="eventLocation"
-                            className="phoneCallModal__window--radio"
-                            value="phone call: (Invitee should call me)"
-                        />
-                        <Typography className="phoneCallModal__window--radio--label" variant="body2">
-                            My invitee should call me<br/>
-                            <span className="phoneCallModal__window--radio--label--span">Calendly will provide your number after the call has been scheduled.</span>
-                        </Typography>
-                    </Box>
-                </Box>
-            </Box>
-            <Box className="phoneCallModal__window--footer">
-                <Button 
-                    onClick={handleLocationUpdate}
-                    className="phoneCallModal__window--btn phoneCallModal__window--btn--update"
-                >Update</Button>
-                <Button 
-                    onClick={handlePhoneCallModal}
-                    className="phoneCallModal__window--btn phoneCallModal__window--btn--cancel"
-                >Cancel</Button>
-            </Box>
-        </Box>
-    </Box>
-);
